// src/app/models/item.ts
export interface Item {
    id?: string;
    title?: string;
    description?: string;
    
    date?: string;
    newCases?: string;
    newDeaths?: string;
    hospitalEX?: string;
    recovered?: string;
    totalDeath?: string;
    totalCases?: string;
    
    name?: string;
    password?: string;

   }